[#54] refactor: parser 구조 변경
