+++ src/main/java/org/unilab/improfessorbe/domain/problem/dto/ConceptExtractionResult.java
@@ -1,4 +1,4 @@
-package org.unilab.improfessorbe.domain.parse.dto;
+package org.unilab.improfessorbe.domain.problem.dto;
 
 import java.util.Collections;
 import java.util.List;
