+++ src/main/java/org/unilab/improfessorbe/domain/problem/parser/FileLogUtil.java
@@ -1,4 +1,4 @@
-package org.unilab.improfessorbe.domain.parse.util;
+package org.unilab.improfessorbe.domain.problem.parser;
 
 import java.util.List;
