+++ src/main/java/org/unilab/improfessorbe/domain/problem/service/ProblemService.java
@@ -11,12 +11,12 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import org.unilab.improfessorbe.domain.parse.dto.ConceptExtractionResult;
-import org.unilab.improfessorbe.domain.parse.input.service.ConceptExtractorOurService;
+import org.unilab.improfessorbe.domain.problem.dto.ConceptExtractionResult;
+import org.unilab.improfessorbe.domain.problem.service.input.ConceptExtractorOurService;
 //import org.unilab.improfessorbe.domain.parse.input.service.ConceptExtractorService;
-import org.unilab.improfessorbe.domain.parse.input.service.FileParseService;
-import org.unilab.improfessorbe.domain.parse.output.PdfExportService;
-import org.unilab.improfessorbe.domain.parse.output.ProblemTextParser;
+import org.unilab.improfessorbe.domain.problem.service.input.FileParseService;
+import org.unilab.improfessorbe.domain.problem.service.output.PdfExportService;
+import org.unilab.improfessorbe.domain.problem.service.output.ProblemTextParser;
 import org.unilab.improfessorbe.domain.problem.dto.CachedProblemDto;
 import org.unilab.improfessorbe.domain.problem.dto.ProblemDownloadResponse;
 import org.unilab.improfessorbe.domain.problem.dto.ProblemGenerationResponse;
