+++ src/main/java/org/unilab/improfessorbe/domain/problem/service/output/ProblemTextParser.java
@@ -1,4 +1,4 @@
-package org.unilab.improfessorbe.domain.parse.output;
+package org.unilab.improfessorbe.domain.problem.service.output;
 
 import java.util.ArrayList;
 import java.util.List;
